fix: Pass the mountpoint target user to storages without owner
Storages that do not have a dedicated owner (e.g. groupfolders, external

storages) currently always assume the current session user as the owner.

This leads to several issues when there is no user session but a node is

obtained through a user folder.

In order to have the correct user available we need to pass the user

that is used to setup a mountpoint along to the storage layer as we

generally assume that an owner is available for those.

Signed-off-by: Julius Härtl <jus@bitgrid.net>

[skip ci]